Update search-ui-app-search-connector to export requestAdapters
** Changes **
In the `search-ui-app-search-connector` package, the requestAdapters
file is responsible for adapting the SearchUI search state into
AppSearch request state. This state adaptation is also required
for export functionality. In the past, the request adapter
was simply copied and pasted in the Beacon App. But this update
makes this request adapter accessible through normal function
import